fix(pipeline): actually using correct bumps
